refactor: move entities of message domain into separate file
