+++ src/messages/mod.rs
@@ -1,2 +1,3 @@
 pub mod database;
+pub mod entities;
 pub mod service;
